Refactor portfolio manager package
